requirements.txt
# dispatch
dill

# tracker
tinydb
gitpython

# misc
wrapt
attrs 
